requirements.txt
customtkinter==5.2.2
ffmpeg_python==0.2.0
Pillow==11.1.0
python-dotenv==1.0.1
pytubefix==8.12.2
Requests==2.32.3
